+++ src-tauri/src/features/system/commands/prompt_assembly.rs
@@ -551,6 +551,7 @@ mod prompt_assembly_tests {
             pending_chat_result_senders: Arc::new(Mutex::new(HashMap::new())),
             pending_chat_delta_channels: Arc::new(Mutex::new(HashMap::new())),
             active_chat_view_bindings: Arc::new(Mutex::new(HashMap::new())),
+            conversation_list_activity_marks: Arc::new(Mutex::new(HashMap::new())),
             dequeue_lock: Arc::new(Mutex::new(())),
             delegate_runtime_threads: Arc::new(Mutex::new(HashMap::new())),
             delegate_recent_threads: Arc::new(Mutex::new(VecDeque::new())),